move growing ingest bin code into lib
